task5

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -1,8 +1,8 @@
 package lang.print.gaps.task5;
 
 public class NumberDivider {
-    void divide(int toBeDivided) {
-    int x = toBeDivided/5;
+    void divide(double toBeDivided) {
+    double x = toBeDivided/5;
     System.out.println (x);
     }
 }
